backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
paddleocr==2.7.3
paddlepaddle==3.2.2
pillow==10.1.0
numpy==1.26.0
torch==2.5.1
transformers==4.57.0
sentencepiece==0.2.0
protobuf==3.20.3
accelerate==0.25.0

